Fix sys.platform when cross-compiling with emscripten (#14888)
This is a workaround for correctly detecting the platform when building
mypy with mypyc on the Emscripten platform. This fixes
https://github.com/mypyc/mypy_mypyc-wheels/issues/62. `pyodide build`
overrides `sysconfig` for the host based on the target (confirmed on the
pyodide matrix, [see also where the code is actually
changed](https://github.com/pyodide/pyodide/blob/e835bf05ff4aa463024aaeb9689ae70ea5771314/pyodide-build/pyodide_build/pypabuild.py#L43-L50)).

This should only change things when checking/building for the emscripten
platform. There isn't really a cleaner workaround that I can think of
unfortunately, the main issue is cross compiling is tricky with
setuptools.