Don't type check most function bodies if ignoring errors (#14150)
If errors are ignored, type checking function bodies often can have no
effect. Remove function bodies after parsing to speed up type checking.

Methods that define attributes have an externally visible effect even if
errors are ignored. The body of any method that assigns to any attribute
is preserved to deal with this (even if it doesn't actually define a new
attribute). Most methods don't assign to an attribute, so stripping
bodies is still effective for methods.

There are a couple of additional interesting things in the
implementation:

1. We need to know whether an abstract method has a trivial body (e.g.
just `...`) to check `super()` method calls. The approach here is to
preserve such trivial bodies and treat them differently from no body at
all.
2. Stubgen analyzes the bodies of functions to e.g. infer some return
types. As a workaround, explicitly preserve full ASTs when using
stubgen.

The main benefit is faster type checking when using installed packages
with inline type information (PEP 561). Errors are ignored in this case,
and it's common to have a large number of third-party code to type
check. For example, a self check (code under `mypy/`) is now about **20%
faster**, with a compiled mypy on Python 3.11.

Another, more subtle benefit is improved reliability. A third-party
library may have some code that triggers a mypy crash or an invalid
blocking error. If bodies are stripped, often the error will no longer
be triggered, since the amount code to type check is much lower.

---------

Co-authored-by: Shantanu <12621235+hauntsaninja@users.noreply.github.com>