[PEP646] Add tests/fixes for callable star args behavior (#15069)
This adds tests for mixing star args with typevar tuples from PEP646. In
order to do this we have to handle an additional case in applytype where
the result of expanding an unpack is a homogenous tuple rather than a
list of types.

This PR also includes (partially by accident) a fix to the empty case
for type analysis. After re-reading the PEP we discover that the empty
case is meant to be inferred as a Tuple[Any, ...] rather than an empty
tuple and must fix that behavior as well as some of the previous test
cases that assumed that it meant an empty tuple.

When we actually call `target(*args)` from the testcase in `call` we
expose a crash in mypy. This will be handled by a subsequent PR.