Fix some TypeVarTuple Callable edge cases.
Previously we mostly only tested substituting in concrete values to
Callables (see tests in PR). Now we test in substituting Unpack[Args],
etc. Prefices and suffices are not implemented yet because they seem
very nontrivial to get working.

We also test what happens when the `submit` function in the test case
is defined using ParamSpec but we pass in an Unpack through *args, and
we fix a crash related to that scenario.