[mypyc] Remove unused labels and gotos:
This PR removes labels which are never jumped to, as well as removing gotos
that jump directly to the next label (and removes the now redundant label).
I had a small issue with the `GetAttr` op code as it would insert C code after
the end of the branch which could not be detected at the IR level, as seen in
https://github.com/mypyc/mypyc/issues/600#issuecomment-1319530780 .

Here are some basic stats on the file sizes before and after this PR:

Building with `MYPY_OPT_LEVEL=0`, `MYPY_DEBUG_LEVEL=1`:

| Branch   | `.so` size      | `.c` size       | `.c` lines      |
|----------|-----------------|-----------------|-----------------|
| `master` | 43.9 MB         | 79.7 MB         | 2290675         |
| This PR  | 43.6 MB (-0.6%) | 78.6 MB (-1.4%) | 2179967 (-4.8%) |

Building with `MYPY_OPT_LEVEL=3`, `MYPY_DEBUG_LEVEL=0`:

| Branch   | `.so` size | `.c` size | `.c` lines |
|----------|------------|-----------|------------|
| `master` | 26.5 MB    | 79.7 MB   | 2291316    |
| This PR  | 26.5 MB    | 78.7 MB   | 2179701    |

(I don't know why the number of lines changes between optimization levels.
It's only a couple hundred lines, perhaps Mypyc is emitting different code
depending on optimization level?)

Unoptimized builds seem to benefit the most, while optimized builds look
prety much the same. I didn't check to see if they are *exactly* the same
binary, but they are probably close if not the same. The biggest win is the
drop in the number of lines C code being generated.

Closes https://github.com/mypyc/mypyc/issues/600