Pass log-file and timeout args to daemon as proper args (#15227)
Currently, when starting the mypy daemon, we're passing the timeout and
log_file as part of the pickled options. This can cause problems if the
daemon has issues with the pickling. In this case we won't be able to
log what the problem was.
Let's pass the log_file and timeout as first class args. This also has
the advantage that we can now start the daemon in the foreground with
these args from the command line in a human writeable way.