testcheck: sort inline assertions (#15290)
The order in which inline error assertions are collected into the
`TestCase.output` (the "expected") can differ from the actual errors'
order. Specifically, it's the order of the files themselves (as they're
traversed during build), rather than individual lines.

To allow inline assertions across multiple modules, we can sort the
collected errors to match the actual errors (only when it concerns
module order; within the modules the order remains stable).