Inline math literals:
This PR inlines math literals such as `math.pi` and `math.e`. Previously
Using `math.pi` would emit a getattr, unbox, error check, then box op code,
but now it just boxes the value directly.

This PR is an improvement, but there are a few things that could be improved
further:

* `math` is still being imported even if you only imported literal values such
  as `pi` (which is now inlined and doesn't need to be imported).

* We are essentially boxing a literal which should be a `CPyLit_Float` value
  instead. I looked into this, but AFAICT there is no "float object" primitve,
  meaning Mypyc will emit the same amount of code as before (minus the getattr)
  which is not ideal. Is there any way to indicate in the IR that a
  `CPyLit_Float` value is a boxed float as opposed to an unboxed float
  (`double`)?

Also, I added a test to ensure that the math literals are inlined properly and
have the same value, but the test is failing saying that `pi` does not exist
on module `math`. There are no errors when running the same file through mypy,
so I don't know why it is failing. I checked, and the issue still seems to be
present on `master`.

Closes https://github.com/mypyc/mypyc/issues/985