Revert "Add support for `attrs.fields` (#15021)"
This reverts commit 391ed853f7857222d7eb209ecdc48f0a130f70d7.

The use of a Union item type for `attr.Attribute` would likely be
more useful to callsites of `attr.fields` - otherwise this currently
results in inferring `builtins.object` in many cases where we treat
the fields as a variable-length tuple, rather than a fixed-length
one. Applications of the fixed-length tuple seem limited since you
would need to access it by index.

@Tinche as the original author of the commit.