Remove special handling of `kw_only` in `dataclass` plugin serialization
It is not needed, because it is always serialized as `bool` and is always present.

It was added a long time ago in https://github.com/python/mypy/pull/10867 as a compat layer. But, since then we've added at least one `is_neither_frozen_nor_nonfrozen` attribute that is serialized / deserialized in a common way.

So, let's remove this hack for good.