Remove `py` and `pytest-forked` test dependencies (#15501)
We don't use either of these in our own code at all. Neither of them
should be needed in 2023.

py used to be a dependency of pytest but no longer is. It was added to
test-requirements.txt 6 years ago in order to pin a specific version to
fix an issue where some tests were failing:
https://github.com/python/mypy/commit/5168c2580ff81ba114776936bdeaf5d81812c34a
and
https://github.com/python/mypy/commit/ad6dd4ed6923a5eb6227fc9bfcb8e30a5e0fde53.

pytest-forked used to be a dependency of pytest-xdist, but no longer is.
It was added to test-requirements.txt` 4 years ago in order to pin a
specific version to fix builds on macOS:
https://github.com/python/mypy/commit/e006b94e097184451e6baf9b9c55e56176728ab8.