[mypyc] Support the i16 native integer type (#15464)
The `i16` type behaves the same as `i64` and `i32`, but is obviously
smaller. The PR is big, but it's mostly because of test cases, which are
adapted from `i32` test cases.

Also fix error handling in unboxing of native int and float types (i.e.
all types with overlapping error values).