When sync-typeshed fails to apply CP allow manual user merges (#15591)
If misc/sync-typeshed.py fails to apply a cherry pick, it just fails.
Let's try to give the user a chance to manually merge the CP and
continue with the script.
This should block for user input only in cases where stdin is a tty. So
automation should continue failing. (but the only way to test is by
running it)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Alex Waygood <Alex.Waygood@Gmail.com>