[mypyc] Fix multiple inheritance with a protocol on Python 3.12 (#15572)
Fix the testProtocol test case in mypyc/test-data/run-classes.test.

There was a comment indicating that the code that causes the test to
fail on 3.12 is required Python versions older than 3.7, so let's remove
it. We only support 3.8 and later these days.

Work on https://github.com/mypyc/mypyc/issues/995.