New type inference: add support for upper bounds and values (#15813)
This is a third PR in series following
https://github.com/python/mypy/pull/15287 and
https://github.com/python/mypy/pull/15754. This one is quite simple: I
just add basic support for polymorphic inference involving type
variables with upper bounds and values. A complete support would be
quite complicated, and it will be a corner case to already rare
situation. Finally, it is written in a way that is easy to tune in the
future.

I also use this PR to add some unit tests for all three PRs so far,
other two PRs only added integration tests (and I clean up existing unit
tests as well).