[dmypy] special case stdout and stderr in show_stats too
When running dmypy, the communication between client and server is via
JSON. The JSON contains the keys "out" and "err" for the actual result
of "check" command, and "stdout" and "stderr" for the any other stdout
and stderr text.
show_stats is shown when running with --verbose. It's meant to show
other keys in the JSON response, like python version or time taken.
It already had some special casing to only show 1 line of text for "out"
and "err". Let's add "stdout" and "stderr" to the special casing as
well. Also, let's show the remaining number of characters as well.