Document and rename overlap-overload error code
A new error code was introduced in
https://github.com/python/mypy/pull/16061

As per https://github.com/python/mypy/pull/16068, we didn't previously
run doc builds on changes to errorcodes.py, causing tests to fail on
master when this was merged.

Renaming the code as per:
https://github.com/python/mypy/pull/16061#issuecomment-1710613890
All type ignores should be unsafe, so we should save the unsafe
adjective for things that are really unsafe. As it stands, there are
many cases where overloads overlap somewhat benignly.