Better support for variadic calls and indexing (#16131)
This improves support for two features that were supported but only
partially: variadic calls, and variadic indexing. Some notes:
* I did not add dedicated support for slicing of tuples with homogeneous
variadic items (except for cases covered by TypeVarTuple support, i.e.
those not involving splitting of variadic item). This is tricky and it
is not clear what cases people actually want. I left a TODO for this.
* I prohibit multiple variadic items in a call expression. Technically,
we can support some situations involving these, but this is tricky, and
prohibiting this would be in the "spirit" of the PEP, IMO.
* I may have still missed some cases for the calls, since there are so
many options. If you have ideas for additional test cases, please let me
know.
* It was necessary to fix overload ambiguity logic to make some tests
pass. This goes beyond TypeVarTuple support, but I think this is a
correct change.